Move SidebarNodes into own file (FileTree)
